Remove unused utils and change float to double

+++ app/src/main/java/io/github/wulkanowy/data/db/entities/Grade.kt
@@ -20,7 +20,7 @@ data class Grade(
 
         var value: Int,
 
-        var modifier: Float,
+        var modifier: Double,
 
         var comment: String,
